Add support for installing Allihoopa in drop flow (#12)
After a piece has been dropped the user can choose to view the piece.
The SDK now tries to open the piece in the Allihoopa app, if it's
installed. If it isn't installed on the current device the user is shown
a modal view where she can decide to install Allihoopa or open the piece
in Safari.